# docs/swagger.yaml
basePath: /
definitions:
  authhandler.CheckOTPAndLogInResponse:
    properties:
      accsessToken:
        type: string
      error:
        type: string
      refreshToken:
        type: string
      status:
        type: string
    type: object
  authhandler.SingInByTgResponse:
    properties:
      error:
        type: string
      info:
        type: string
      status:
        type: string
      success:
        type: boolean
    type: object
  authhandler.SingInResponse:
    properties:
      accsessToken:
        type: string
      error:
        type: string
      refreshToken:
        type: string
      status:
        type: string
    type: object
  authhandler.SingUpResponse:
    properties:
      error:
        type: string
      status:
        type: string
      success:
        type: boolean
    type: object
  authhandler.UpdateUserInfoReq:
    properties:
      email:
        type: string
      is_admin:
        type: boolean
      name:
        type: string
      tg_link:
        type: string
    type: object
  authhandler.UpdateUserInfoResponse:
    properties:
      error:
        type: string
      status:
        type: string
      success:
        type: boolean
    type: object
...
  response.Response:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
  ssomodels.CheckOTPAndLogIn:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
...
  ssomodels.LogIn:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  ssomodels.LogInViaTg:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  ssomodels.RegisterUser:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
host: localhost:8000
info:
  contact:
    name: API Support
  description: The project is only in its initial stages.
  title: Learning Platform API
  version: 0.1.0
paths:
  /check_otp:
    post:
      consumes:
      - application/json
      description: This endpoint allows users to sign in using their email and sends
        OTP code to chat.
      parameters:
      - description: Sign-in parameters
        in: body
        name: ssomodels.CheckOTPAndLogIn
        required: true
        schema:
          $ref: '#/definitions/ssomodels.CheckOTPAndLogIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authhandler.CheckOTPAndLogInResponse'
        "400":
          description: Invalid data in the request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Login by telegram bot
      tags:
      - auth
...
  /profile/update_info:
    patch:
      consumes:
      - application/json
      description: This endpoint allow users to change their profile.
      parameters:
      - description: Sign-in parameters
        in: body
        name: authhandler.UpdateUserInfoReq
        required: true
        schema:
          $ref: '#/definitions/authhandler.UpdateUserInfoReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authhandler.UpdateUserInfoResponse'
        "400":
          description: Invalid data in the request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Change self user info
      tags:
      - auth
  /sing_in:
    post:
      consumes:
      - application/json
      description: This endpoint allows users to sign in using their email and password.
      parameters:
      - description: Sign-in parameters
        in: body
        name: ssomodels.LogIn
        required: true
        schema:
          $ref: '#/definitions/ssomodels.LogIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authhandler.SingInResponse'
        "400":
          description: Invalid data in the request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Login
      tags:
      - auth
  /sing_in_by_tg:
    post:
      consumes:
      - application/json
      description: This endpoint allows users to sign in using their email and sends
        OTP code to chat.
      parameters:
      - description: Sign-in parameters
        in: body
        name: ssomodels.LogInViaTg
        required: true
        schema:
          $ref: '#/definitions/ssomodels.LogInViaTg'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authhandler.SingInByTgResponse'
        "400":
          description: Invalid data in the request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Login by telegram bot
      tags:
      - auth
  /sing_up:
    post:
      consumes:
      - application/json
      description: This endpoint allows users to register with an email and password.
      parameters:
      - description: Registration parameters
        in: body
        name: ssomodels.RegisterUser
        required: true
        schema:
          $ref: '#/definitions/ssomodels.RegisterUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/authhandler.SingUpResponse'
        "400":
          description: Invalid data in the request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register a new user
      tags:
      - auth
security:
- ApiKeyAuth: []
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"